requirements.txt
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.2
dj-database-url==1.0.0
gunicorn==20.1.0
wagtail==4.0.1
djangorestframework==3.13.1
django-colorfield==0.7.2
